+++ packages/desktop/src/renderer/components/Channel/CreateChannel/CreateChannel.stories.tsx
@@ -3,6 +3,7 @@ import { ComponentStory, ComponentMeta } from '@storybook/react'
 
 import { withTheme } from '../../../storybook/decorators'
 import CreateChannelComponent, { CreateChannelProps } from './CreateChannelComponent'
+import { defaultLogger } from '../../../logger'
 
 const Template: ComponentStory<typeof CreateChannelComponent> = args => {
   return <CreateChannelComponent {...args} />
@@ -13,7 +14,7 @@ export const Component = Template.bind({})
 const args: CreateChannelProps = {
   open: true,
   createChannel: function (name: string): void {
-    console.log('creating channel: ', name)
+    defaultLogger.info('creating channel: ', name)
   },
   handleClose: function (): void {},
   clearErrorsDispatch: function (): void {},