+++ packages/desktop/src/renderer/components/CreateUsername/CreateUsernameComponent.tsx
@@ -18,6 +18,7 @@ import { TextInput } from '../../forms/components/textInput'
 import { userNameField } from '../../forms/fields/createUserFields'
 
 import { parseName } from '@quiet/common'
+import { defaultLogger } from '../../logger'
 
 const PREFIX = 'CreateUsernameComponent-'
 
@@ -174,7 +175,7 @@ export const CreateUsernameComponent: React.FC<CreateUsernameComponentProps> = (
   const onSubmit = useCallback(
     (values: CreateUserValues) => {
       if (errors.userName) {
-        console.error('Cannot submit form with errors')
+        defaultLogger.error('Cannot submit form with errors')
         return
       }