+++ packages/state-manager/src/sagas/appConnection/connection.selectors.ts
@@ -11,6 +11,9 @@ import { areMessagesLoaded, areChannelsLoaded } from '../publicChannels/publicCh
 import { identitySelectors } from '../identity/identity.selectors'
 import { communitiesSelectors } from '../communities/communities.selectors'
 import { InvitationDataVersion } from '@quiet/types'
+import createLogger from '../../utils/logger'
+
+const logger = createLogger('connection')
 
 const connectionSlice: CreatedSelectors[StoreKeys.Connection] = (state: StoreState) => state[StoreKeys.Connection]
 
@@ -81,7 +84,9 @@ export const isJoiningCompleted = createSelector(
   areChannelsLoaded,
   areCertificatesLoaded,
   (isCommunity, areMessages, areChannels, areCertificates) => {
-    console.log({ isCommunity, areMessages, areChannels, areCertificates })
+    logger.info(
+      `Checking if joining is complete: ${(JSON.stringify({ isCommunity, areMessages, areChannels, areCertificates }), null, 2)}`
+    )
     return isCommunity && areMessages && areChannels && areCertificates
   }
 )