+++ packages/state-manager/src/sagas/communities/communities.selectors.ts
@@ -3,6 +3,9 @@ import { createSelector } from 'reselect'
 import { communitiesAdapter } from './communities.adapter'
 import { type CreatedSelectors, type StoreState } from '../store.types'
 import { CertFieldsTypes, getCertFieldValue, parseCertificate } from '@quiet/identity'
+import createLogger from '../../utils/logger'
+
+const logger = createLogger('communities')
 
 // Workaround for "The inferred type of 'communitiesSelectors' cannot be named without a reference to
 // 'packages/identity/node_modules/pkijs/build'. This is likely not portable. A type annotation is necessary."
@@ -30,6 +33,10 @@ export const currentCommunity = createSelector(communitiesSlice, selectEntities,
 
 export const currentCommunityId = createSelector(communitiesSlice, reducerState => {
   return reducerState.currentCommunity
+})
+
+export const peerList = createSelector(currentCommunity, currentCommunity => {
+  return currentCommunity?.peerList
 })
 
 export const invitationCodes = createSelector(communitiesSlice, reducerState => {
@@ -60,7 +67,7 @@ export const ownerNickname = createSelector(ownerCertificate, ownerCertificate =
   const nickname = getCertFieldValue(parsedCert, CertFieldsTypes.nickName)
 
   if (!nickname) {
-    console.error('Could not retrieve owner nickname from certificate')
+    logger.error('Could not retrieve owner nickname from certificate')
   }
 
   return nickname
@@ -78,4 +85,5 @@ export const communitiesSelectors = {
   ownerCertificate,
   ownerNickname,
   psk,
+  peerList,
 }