+++ packages/state-manager/src/sagas/users/userProfile/saveUserProfile.saga.ts
@@ -11,11 +11,11 @@ import { fileToBase64String } from '@quiet/common'
 
 import { config } from '../../users/const/certFieldTypes'
 import { identitySelectors } from '../../identity/identity.selectors'
-import logger from '../../../utils/logger'
+import createLogger from '../../../utils/logger'
 import { usersActions } from '../users.slice'
 import { type Socket, applyEmitParams } from '../../../types'
 
-const log = logger('saveUserProfileSaga')
+const logger = createLogger('saveUserProfileSaga')
 
 export function* saveUserProfileSaga(socket: Socket, action: PayloadAction<{ photo?: File }>): Generator {
   const identity = yield* select(identitySelectors.currentIdentity)
@@ -28,7 +28,7 @@ export function* saveUserProfileSaga(socket: Socket, action: PayloadAction<{ pho
   try {
     base64EncodedPhoto = yield* call(fileToBase64String, action.payload.photo)
   } catch (err) {
-    log.error('Failed to base64 encode profile photo', err)
+    logger.error('Failed to base64 encode profile photo', err)
     return
   }
 
@@ -47,7 +47,7 @@ export function* saveUserProfileSaga(socket: Socket, action: PayloadAction<{ pho
     pubKey,
   }
 
-  console.log('Saving user profile', userProfile)
+  logger.info(`Saving user profile: ${userProfile}`)
 
   yield* apply(socket, socket.emit, applyEmitParams(SocketActionTypes.SET_USER_PROFILE, userProfile))
 }