+++ packages/desktop/CHANGELOG.md
@@ -2,7 +2,12 @@
 
 # New features:
 
+* Adds connection status information to messages panel when no peers are connected ([#1706](https://github.com/TryQuiet/quiet/issues/1706))
+
 # Refactorings:
+
+* Updates all logging from renderer to write to the node console
+* Updates usages of `act` in tests to use `@testing-library/react` to avoid errors/warnings
 
 # Fixes: